proj1/requirements.txt
langchain-community>=0.0.38
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0